refactor: create cost-insights-common isomorphic package
This moves most of the cost-insights API types into a separate
isomorphic @backstage/plugin-cost-insights-common package. This will
allow these types to be used by a a cost-insights backend or other
product specific modules.
