+++ plugins/cost-insights-common/CHANGELOG.md
@@ -1,0 +1,1 @@
+# @backstage/plugin-cost-insights-common
