+++ plugins/cost-insights-common/README.md
@@ -1,0 +1,8 @@
+# Cost Insights Common
+
+Shared isomorphic code for the cost-insights plugin.
+
+## Links
+
+- [Frontend part of the plugin](https://github.com/backstage/backstage/tree/master/plugins/cost-insights)
+- [The Backstage homepage](https://backstage.io)
