+++ plugins/cost-insights-common/api-report.md
@@ -1,0 +1,185 @@
+## API Report File for "@backstage/plugin-cost-insights-common"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+// Warning: (ae-missing-release-tag) "ChangeStatistic" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export interface ChangeStatistic {
+  // (undocumented)
+  amount: number;
+  // (undocumented)
+  ratio?: number;
+}
+
+// Warning: (ae-missing-release-tag) "ChartData" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export type ChartData = {
+  date: number;
+  trend: number;
+  dailyCost: number;
+  [key: string]: number;
+};
+
+// Warning: (ae-missing-release-tag) "Cost" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export interface Cost {
+  // (undocumented)
+  aggregation: DateAggregation[];
+  // (undocumented)
+  change?: ChangeStatistic;
+  // (undocumented)
+  groupedCosts?: Record<string, Cost[]>;
+  // (undocumented)
+  id: string;
+  // (undocumented)
+  trendline?: Trendline;
+}
+
+// Warning: (ae-missing-release-tag) "Currency" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export interface Currency {
+  // (undocumented)
+  kind: string | null;
+  // (undocumented)
+  label: string;
+  // (undocumented)
+  prefix?: string;
+  // (undocumented)
+  rate?: number;
+  // (undocumented)
+  unit: string;
+}
+
+// Warning: (ae-missing-release-tag) "DateAggregation" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export type DateAggregation = {
+  date: string;
+  amount: number;
+};
+
+// Warning: (ae-missing-release-tag) "Duration" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public
+export enum Duration {
+  // (undocumented)
+  P30D = 'P30D',
+  // (undocumented)
+  P3M = 'P3M',
+  // (undocumented)
+  P7D = 'P7D',
+  // (undocumented)
+  P90D = 'P90D',
+}
+
+// Warning: (ae-missing-release-tag) "Entity" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export interface Entity {
+  // (undocumented)
+  aggregation: [number, number];
+  // (undocumented)
+  change: ChangeStatistic;
+  // (undocumented)
+  entities: Record<string, Entity[]>;
+  // (undocumented)
+  id: Maybe<string>;
+}
+
+// Warning: (ae-missing-release-tag) "Group" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export type Group = {
+  id: string;
+};
+
+// Warning: (ae-missing-release-tag) "Maybe" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export type Maybe<T> = T | null;
+
+// Warning: (ae-missing-release-tag) "Metric" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export type Metric = {
+  kind: string;
+  name: string;
+  default: boolean;
+};
+
+// Warning: (ae-missing-release-tag) "MetricData" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export interface MetricData {
+  // (undocumented)
+  aggregation: DateAggregation[];
+  // (undocumented)
+  change: ChangeStatistic;
+  // (undocumented)
+  format: 'number' | 'currency';
+  // (undocumented)
+  id: string;
+}
+
+// Warning: (ae-missing-release-tag) "PageFilters" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export interface PageFilters {
+  // (undocumented)
+  duration: Duration;
+  // (undocumented)
+  group: Maybe<string>;
+  // (undocumented)
+  metric: string | null;
+  // (undocumented)
+  project: Maybe<string>;
+}
+
+// Warning: (ae-missing-release-tag) "Product" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export interface Product {
+  // (undocumented)
+  kind: string;
+  // (undocumented)
+  name: string;
+}
+
+// Warning: (ae-missing-release-tag) "ProductFilters" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export type ProductFilters = Array<ProductPeriod>;
+
+// Warning: (ae-missing-release-tag) "ProductPeriod" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export interface ProductPeriod {
+  // (undocumented)
+  duration: Duration;
+  // (undocumented)
+  productType: string;
+}
+
+// Warning: (ae-missing-release-tag) "Project" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export interface Project {
+  // (undocumented)
+  id: string;
+  // (undocumented)
+  name?: string;
+}
+
+// Warning: (ae-missing-release-tag) "Trendline" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export type Trendline = {
+  slope: number;
+  intercept: number;
+};
+```
