+++ plugins/cost-insights-common/src/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Backstage Authors
+ * Copyright 2022 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-export enum CurrencyType {
-  USD = 'USD',
-  CarbonOffsetTons = 'CARBON_OFFSET_TONS',
-  Beers = 'BEERS',
-  IceCream = 'PINTS_OF_ICE_CREAM',
-}
+/**
+ * Provides shared objects useful for interacting with the cost-insights plugin.
+ *
+ * @packageDocumentation
+ */
+
+export * from './types';
