+++ plugins/cost-insights-common/src/types/ChangeStatistic.ts
@@ -14,9 +14,11 @@
  * limitations under the License.
  */
 
-export enum CurrencyType {
-  USD = 'USD',
-  CarbonOffsetTons = 'CARBON_OFFSET_TONS',
-  Beers = 'BEERS',
-  IceCream = 'PINTS_OF_ICE_CREAM',
+export interface ChangeStatistic {
+  // The ratio of change from one duration to another, expressed as: (newSum - oldSum) / oldSum
+  // If a ratio cannot be calculated - such as when a new or old sum is zero,
+  // the ratio can be omitted and where applicable, ∞ or -∞ will display based on amount.
+  ratio?: number;
+  // The actual USD change between time periods (can be negative if costs decreased)
+  amount: number;
 }
