+++ plugins/cost-insights-common/src/types/ChartData.ts
