+++ plugins/cost-insights-common/src/types/Duration.ts
@@ -14,9 +14,15 @@
  * limitations under the License.
  */
 
-export enum CurrencyType {
-  USD = 'USD',
-  CarbonOffsetTons = 'CARBON_OFFSET_TONS',
-  Beers = 'BEERS',
-  IceCream = 'PINTS_OF_ICE_CREAM',
+/**
+ * Time periods for cost comparison; slight abuse of ISO 8601 periods. We take P3M to mean
+ * 'last completed quarter', and P30D/P90D to be '[month|quarter] relative to today'. So if
+ * it's September 15, P3M represents costs for Q2 and P30D represents August 16 -
+ * September 15.
+ */
+export enum Duration {
+  P7D = 'P7D',
+  P30D = 'P30D',
+  P90D = 'P90D',
+  P3M = 'P3M',
 }
