+++ plugins/cost-insights-common/src/types/Entity.ts
