+++ plugins/cost-insights-common/src/types/Product.ts
