+++ plugins/cost-insights-common/src/types/Trendline.ts
