+++ plugins/cost-insights-common/src/types/index.ts
@@ -14,9 +14,18 @@
  * limitations under the License.
  */
 
-export enum CurrencyType {
-  USD = 'USD',
-  CarbonOffsetTons = 'CARBON_OFFSET_TONS',
-  Beers = 'BEERS',
-  IceCream = 'PINTS_OF_ICE_CREAM',
-}
+export * from './ChangeStatistic';
+export * from './ChartData';
+export * from './Cost';
+export * from './DateAggregation';
+export * from './Duration';
+export * from './Currency';
+export * from './Entity';
+export * from './Filters';
+export * from './Group';
+export * from './Maybe';
+export * from './MetricData';
+export * from './Metric';
+export * from './Product';
+export * from './Project';
+export * from './Trendline';
