+++ plugins/cost-insights/package.json
@@ -38,6 +38,7 @@
     "@backstage/config": "^1.0.1",
     "@backstage/core-components": "^0.9.6-next.1",
     "@backstage/core-plugin-api": "^1.0.3",
+    "@backstage/plugin-cost-insights-common": "^0.0.1",
     "@backstage/theme": "^0.2.16-next.0",
     "@material-ui/core": "^4.12.2",
     "@material-ui/icons": "^4.9.1",
