+++ plugins/cost-insights/src/types/ChangeStatistic.ts
@@ -14,15 +14,6 @@
  * limitations under the License.
  */
 
-export interface ChangeStatistic {
-  // The ratio of change from one duration to another, expressed as: (newSum - oldSum) / oldSum
-  // If a ratio cannot be calculated - such as when a new or old sum is zero,
-  // the ratio can be omitted and where applicable, ∞ or -∞ will display based on amount.
-  ratio?: number;
-  // The actual USD change between time periods (can be negative if costs decreased)
-  amount: number;
-}
-
 export const EngineerThreshold = 0.5;
 
 export enum ChangeThreshold {
